qtenv: refactor: Store a map from message pointers to animations that animate them
To speed up the removal of animations when messages are deleted.
This way, we know where to look for the affected animations, indexing
into this map with only the two messages (original in the module and private
duplicate in the LogBuffer) have to be looked up, instead of iterating
on the entire `animations` map, recursing thrugh the animation tree.