Revert "sim: cDisplayString: Be more strict about backslash escape sequences"
This reverts commit 0f3b7446b57511c917b8f87ca755a65afdecd6ce.

The reverted commit turned out to be a bad idea because in this stage,
the "\" escape sequence interpretation doesn't add any special meaning
(unlike during NED parsing of string literals, with "\n" and "\t" for
example), so it doesn't make sense to restrict what can follow backslash
characters. Here, backslashes simply take away any special meaning the
next character may or may not have - so no need to be strict about it.